requirements.txt
solace-pubsubplus
requests
flask